Protecting Routes & Profile page
Protecting routes depending if an user is authenticated, is no possible to see login and signUp pages, on the other side if is no user authenticated, is no possible to see the dashboard.
Profile page created, many components added for the profile, and a second routing for the sections in the profile added.